Refactor check to support new implementation
