+++ win32_event_log/datadog_checks/win32_event_log/__init__.py
@@ -2,6 +2,6 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .__about__ import __version__
-from .win32_event_log import Win32EventLogWMI
+from .win32_event_log import Win32EventLogCheck
 
-__all__ = ['__version__', 'Win32EventLogWMI']
+__all__ = ['__version__', 'Win32EventLogCheck']
