+++ win32_event_log/tests/legacy/common.py
