+++ win32_event_log/tests/legacy/test_win32_integration.py
@@ -5,7 +5,7 @@ import platform
 
 import pytest
 
-from datadog_checks.win32_event_log import Win32EventLogWMI
+from datadog_checks.win32_event_log import Win32EventLogCheck
 
 from . import common
 
@@ -13,6 +13,6 @@ from . import common
 @pytest.mark.integration
 @pytest.mark.skipif(platform.system() != 'Windows', reason="Test only valid on Windows")
 def test_basic_check(aggregator):
-    check = Win32EventLogWMI('win32_event_log', {}, {}, [common.INSTANCE])
+    check = Win32EventLogCheck('win32_event_log', {}, [common.INSTANCE])
     check.check(common.INSTANCE)  # First run just initialises timestamp
     check.check(common.INSTANCE)
